requirements.txt
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.4.0
azt==0.0.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coverage==6.4
Cython==0.29.30
distlib==0.3.4
executing==0.8.3
filelock==3.7.0
idna==3.3
iniconfig==1.1.1
littleutils==0.2.2
mysql-connector-python==8.0.29
packaging==21.3
pluggy==1.0.0
protobuf==4.21.1
py==1.11.0
pyfiglet==0.8.post1
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-mock==3.7.0
pytz==2022.1
requests==2.27.1
requests-mock==1.9.3
schedule==1.1.0
six==1.16.0
sorcery==0.2.2
termcolor==1.1.0
tomli==2.0.1
urllib3==1.26.9
wrapt==1.14.1
